+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+* Added: `Chat` component
+* Added: `Tile` component
+* Added: `NameValueList.{List,Name,Value}` utility components
+
 ## 8.9.0
 
 * Changed: Add optional title menu action to `ObjectHome`